+++ common/src/main/java/net/spell_engine/SpellEngineMod.java
@@ -8,13 +8,13 @@ import net.minecraft.advancement.criterion.Criteria;
 import net.minecraft.client.resource.language.I18n;
 import net.minecraft.enchantment.Enchantments;
 import net.minecraft.entity.EntityType;
+import net.minecraft.entity.EquipmentSlot;
 import net.minecraft.entity.attribute.EntityAttributes;
+import net.minecraft.item.ArmorItem;
 import net.minecraft.item.BlockItem;
 import net.minecraft.item.ItemGroup;
-import net.minecraft.util.Identifier;
 import net.minecraft.util.registry.Registry;
 import net.spell_engine.api.enchantment.Enchantments_SpellEngine;
-import net.spell_engine.api.item.trinket.SpellBooks;
 import net.spell_engine.api.item.weapon.StaffItem;
 import net.spell_engine.config.EnchantmentsConfig;
 import net.spell_engine.config.ServerConfig;
@@ -27,6 +27,7 @@ import net.spell_engine.network.ServerNetwork;
 import net.spell_engine.particle.Particles;
 import net.spell_engine.spellbinding.*;
 import net.spell_power.api.enchantment.EnchantmentRestriction;
+import net.spell_power.api.enchantment.Enchantments_SpellPower;
 import net.spell_power.api.enchantment.SpellPowerEnchanting;
 import net.tinyconfig.ConfigManager;
 
@@ -61,6 +62,17 @@ public class SpellEngineMod {
         EnchantmentRestriction.permit(Enchantments.KNOCKBACK, itemStack -> itemStack.getItem() instanceof StaffItem);
         EnchantmentRestriction.permit(Enchantments.LOOTING, itemStack -> itemStack.getItem() instanceof StaffItem);
         EnchantmentRestriction.permit(Enchantments.FIRE_ASPECT, itemStack -> itemStack.getItem() instanceof StaffItem);
+        EnchantmentRestriction.prohibit(Enchantments_SpellPower.HASTE, itemStack -> {
+            var item = itemStack.getItem();
+            EquipmentSlot slot;
+            if (item instanceof ArmorItem armorItem) {
+                slot = armorItem.getSlotType();
+            } else {
+                slot = EquipmentSlot.MAINHAND;
+            }
+            var empty = SpellPowerEnchanting.relevantSchools(itemStack, slot).isEmpty();
+            return empty;
+        });
 
         SpellPowerEnchanting.allowForWeapon(SpellContainerHelper::hasValidContainer);